requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==3.2.22
django-cors-headers==4.1.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
